Move apiService to services/api
